+++ src/main/java/com/bag/osmanager/service/RelatorioService.java
@@ -16,10 +16,9 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.time.DayOfWeek;
+// ✨ ALTERAÇÃO AQUI: Imports de DayOfWeek, LocalTime, etc., foram removidos pois a lógica saiu daqui
 import java.time.Duration;
 import java.time.LocalDateTime;
-import java.time.LocalTime;
 import java.util.ArrayList; 
 import java.util.Comparator; 
 import java.util.HashMap;
@@ -37,71 +36,8 @@ public class RelatorioService {
 
     private final OrdemServicoRepository ordemServicoRepository;
     
-    // ... (Map de horários e classe Turno permanecem os mesmos) ...
-    private static class Turno {
-        final LocalTime inicio;
-        final LocalTime fim;
-        Turno(LocalTime inicio, LocalTime fim) { this.inicio = inicio; this.fim = fim; }
-        boolean viraDia() { return inicio.isAfter(fim); }
-    }
-
-    private static final Map<String, Map<DayOfWeek, Turno>> horariosMecanicos = new HashMap<>();
-    static {
-        // Horários de ISRAEL
-        Map<DayOfWeek, Turno> horarioIsrael = new HashMap<>();
-        Turno turnoSemanaIsrael = new Turno(LocalTime.of(13, 36), LocalTime.of(22, 10));
-        horarioIsrael.put(DayOfWeek.MONDAY, turnoSemanaIsrael);
-        horarioIsrael.put(DayOfWeek.TUESDAY, turnoSemanaIsrael);
-        horarioIsrael.put(DayOfWeek.WEDNESDAY, turnoSemanaIsrael);
-        horarioIsrael.put(DayOfWeek.THURSDAY, turnoSemanaIsrael);
-        horarioIsrael.put(DayOfWeek.FRIDAY, turnoSemanaIsrael);
-        horarioIsrael.put(DayOfWeek.SATURDAY, new Turno(LocalTime.of(11, 15), LocalTime.of(17, 30)));
-        horariosMecanicos.put("ISRAEL", horarioIsrael);
-
-        // Horários de MAURICIO
-        Map<DayOfWeek, Turno> horarioMauricio = new HashMap<>();
-        Turno turnoSemanaMauricio = new Turno(LocalTime.of(5, 0), LocalTime.of(13, 36));
-        horarioMauricio.put(DayOfWeek.MONDAY, turnoSemanaMauricio);
-        horarioMauricio.put(DayOfWeek.TUESDAY, turnoSemanaMauricio);
-        horarioMauricio.put(DayOfWeek.WEDNESDAY, turnoSemanaMauricio);
-        horarioMauricio.put(DayOfWeek.THURSDAY, turnoSemanaMauricio);
-        horarioMauricio.put(DayOfWeek.FRIDAY, turnoSemanaMauricio);
-        horarioMauricio.put(DayOfWeek.SATURDAY, new Turno(LocalTime.of(5, 0), LocalTime.of(11, 15)));
-        horariosMecanicos.put("MAURICIO", horarioMauricio);
-
-        // Horários de ROGER
-        Map<DayOfWeek, Turno> horarioRoger = new HashMap<>();
-        Turno turnoSemanaRoger = new Turno(LocalTime.of(5, 0), LocalTime.of(13, 36));
-        horarioRoger.put(DayOfWeek.MONDAY, turnoSemanaRoger);
-        horarioRoger.put(DayOfWeek.TUESDAY, turnoSemanaRoger);
-        horarioRoger.put(DayOfWeek.WEDNESDAY, turnoSemanaRoger);
-        horarioRoger.put(DayOfWeek.THURSDAY, turnoSemanaRoger);
-        horarioRoger.put(DayOfWeek.FRIDAY, turnoSemanaRoger);
-        horarioRoger.put(DayOfWeek.SATURDAY, new Turno(LocalTime.of(5, 0), LocalTime.of(11, 15)));
-        horariosMecanicos.put("ROGER", horarioRoger);
-
-        // Horários de ANTONIO
-        Map<DayOfWeek, Turno> horarioAntonio = new HashMap<>();
-        Turno turnoSemanaAntonio = new Turno(LocalTime.of(8, 0), LocalTime.of(17, 0));
-        horarioAntonio.put(DayOfWeek.MONDAY, turnoSemanaAntonio);
-        horarioAntonio.put(DayOfWeek.TUESDAY, turnoSemanaAntonio);
-        horarioAntonio.put(DayOfWeek.WEDNESDAY, turnoSemanaAntonio);
-        horarioAntonio.put(DayOfWeek.THURSDAY, turnoSemanaAntonio);
-        horarioAntonio.put(DayOfWeek.FRIDAY, turnoSemanaAntonio);
-        horarioAntonio.put(DayOfWeek.SATURDAY, new Turno(LocalTime.of(8, 0), LocalTime.of(12, 0)));
-        horariosMecanicos.put("ANTONIO", horarioAntonio);
-        
-        // Horários de CARLOS
-        Map<DayOfWeek, Turno> horarioCarlos = new HashMap<>();
-        Turno turnoSemanaCarlos = new Turno(LocalTime.of(8, 0), LocalTime.of(17, 0)); 
-        horarioCarlos.put(DayOfWeek.MONDAY, turnoSemanaCarlos);
-        horarioCarlos.put(DayOfWeek.TUESDAY, turnoSemanaCarlos);
-        horarioCarlos.put(DayOfWeek.WEDNESDAY, turnoSemanaCarlos);
-        horarioCarlos.put(DayOfWeek.THURSDAY, turnoSemanaCarlos);
-        horarioCarlos.put(DayOfWeek.FRIDAY, turnoSemanaCarlos);
-        horarioCarlos.put(DayOfWeek.SATURDAY, new Turno(LocalTime.of(8, 0), LocalTime.of(12, 0))); 
-        horariosMecanicos.put("CARLOS", horarioCarlos); 
-    }
+    // ✨ ALTERAÇÃO AQUI: A classe interna 'Turno' foi removida.
+    // ✨ ALTERAÇÃO AQUI: O mapa 'horariosMecanicos' foi removido.
 
     /**
      * Agrega todos os dados de relatório para o dashboard do líder em um período.
@@ -126,7 +62,7 @@ public class RelatorioService {
     }
 
     // --- GRÁFICO 1: TEMPO POR MECÂNICO ---
-    // ✨✅ ALTERAÇÃO AQUI: Lógica de cálculo HÍBRIDA (Relatórios + Fallback)
+    // (Esta lógica permanece a mesma, pois ela lê os minutos do AcompanhamentoOS)
     private List<RelatorioTempoMecanicoDTO> gerarRelatorioTempoMecanicos(LocalDateTime inicioPeriodo, LocalDateTime fimPeriodo) {
         
         // 1. Busca OSs que ESTAVAM ATIVAS no período
@@ -158,9 +94,9 @@ public class RelatorioService {
             // Filtra as OSs que foram *concluídas* dentro do período para o total de OS
             long totalOsConcluidas = osDoMecanico.stream()
                 .filter(os -> os.getStatus() == StatusOrdemServico.CONCLUIDA &&
-                               os.getTermino() != null &&
-                               !os.getTermino().isBefore(inicioPeriodo) && 
-                               !os.getTermino().isAfter(fimPeriodo))
+                            os.getTermino() != null &&
+                            !os.getTermino().isBefore(inicioPeriodo) && 
+                            !os.getTermino().isAfter(fimPeriodo))
                 .count();
             
             // Soma os 'minutosTrabalhados' dos Acompanhamentos que estão DENTRO do período
@@ -174,10 +110,10 @@ public class RelatorioService {
                     if (os.getAcompanhamentos() != null && !os.getAcompanhamentos().isEmpty()) {
                         minutosTrabalhadosRegistrados = os.getAcompanhamentos().stream()
                             .filter(acomp -> acomp.getFuncionario() != null && 
-                                             nomeMecanico.equals(acomp.getFuncionario().getNome()) &&
-                                             acomp.getMinutosTrabalhados() != null &&
-                                             !acomp.getDataHora().isBefore(inicioPeriodo) && // Não antes do início
-                                             !acomp.getDataHora().isAfter(fimPeriodo))      // Não depois do fim
+                                            nomeMecanico.equals(acomp.getFuncionario().getNome()) &&
+                                            acomp.getMinutosTrabalhados() != null &&
+                                            !acomp.getDataHora().isBefore(inicioPeriodo) && // Não antes do início
+                                            !acomp.getDataHora().isAfter(fimPeriodo))      // Não depois do fim
                             .mapToDouble(AcompanhamentoOS::getMinutosTrabalhados)
                             .sum();
                         
@@ -186,7 +122,7 @@ public class RelatorioService {
                            usouNovoSistema = true;
                         } else {
                            usouNovoSistema = os.getAcompanhamentos().stream()
-                             .anyMatch(acomp -> acomp.getFuncionario() != null &&
+                                .anyMatch(acomp -> acomp.getFuncionario() != null &&
                                                 nomeMecanico.equals(acomp.getFuncionario().getNome()) &&
                                                 !acomp.getDataHora().isBefore(inicioPeriodo) &&
                                                 !acomp.getDataHora().isAfter(fimPeriodo));
@@ -230,7 +166,6 @@ public class RelatorioService {
     }
 
     // --- GRÁFICO 2: RANKING DE CORRETIVAS POR EQUIPAMENTO ---
-    // (Esta lógica está correta - ela conta OSs CONCLUÍDAS no período)
     private List<RelatorioEquipamentoDTO> gerarRelatorioRankingEquipamentos(LocalDateTime inicioPeriodo, LocalDateTime fimPeriodo) {
         
         List<OrdemServico> osCorretivas = ordemServicoRepository
@@ -256,7 +191,6 @@ public class RelatorioService {
     }
 
     // --- GRÁFICO 3: RANKING DE DOWNTIME POR EQUIPAMENTO ---
-    // ✨✅ ALTERAÇÃO AQUI: Lógica de busca e cálculo atualizada
     private List<RelatorioEquipamentoDTO> gerarRelatorioDowntime(LocalDateTime inicioPeriodo, LocalDateTime fimPeriodo) {
         
         // 1. Busca OSs com downtime que ESTAVAM ATIVAS no período
@@ -307,7 +241,6 @@ public class RelatorioService {
     }
 
     // --- GRÁFICO 4: SAÚDE (PREVENTIVA vs CORRETIVA) ---
-    // (Esta lógica está correta - ela conta OSs CONCLUÍDAS no período)
     private List<RelatorioTipoManutencaoDTO> gerarRelatorioSaude(LocalDateTime inicioPeriodo, LocalDateTime fimPeriodo) {
         
         List<OrdemServico> osConcluidas = ordemServicoRepository
@@ -329,7 +262,6 @@ public class RelatorioService {
     }
 
     // --- GRÁFICO 5: INDICADORES (MTTR / MTBF) ---
-    // ✨✅ ALTERAÇÃO AQUI: Lógica de busca e cálculo atualizada
     private RelatorioIndicadoresDTO gerarRelatorioIndicadores(LocalDateTime inicioPeriodo, LocalDateTime fimPeriodo) {
         
         // 1. Busca todas as OSs de corretiva que tiveram máquina parada e estavam ativas no período
@@ -368,8 +300,8 @@ public class RelatorioService {
         // (Aqui consideramos que uma "falha" é o *início* do downtime)
         long numeroDeFalhasNoPeriodo = falhas.stream()
             .filter(os -> os.getInicioDowntime() != null &&
-                           !os.getInicioDowntime().isBefore(inicioPeriodo) &&
-                           !os.getInicioDowntime().isAfter(fimPeriodo))
+                         !os.getInicioDowntime().isBefore(inicioPeriodo) &&
+                         !os.getInicioDowntime().isAfter(fimPeriodo))
             .count();
 
         if (numeroDeFalhasNoPeriodo == 0) {
@@ -400,64 +332,5 @@ public class RelatorioService {
         return new RelatorioIndicadoresDTO(mttr, mtbf);
     }
 
-
-    // (A função calcularMinutosUteis não é mais usada para o relatório de produtividade)
-    private long calcularMinutosUteis(LocalDateTime inicioOS, LocalDateTime fimOS, String nomeMecanico) {
-        
-        String nomeNormalizado = nomeMecanico.trim().toUpperCase().split(" ")[0]; 
-        Map<DayOfWeek, Turno> horario = horariosMecanicos.get(nomeNormalizado);
-
-        if (horario == null) {
-            String[] nomes = nomeMecanico.trim().toUpperCase().split(" ");
-            if (nomes.length > 1) {
-                horario = horariosMecanicos.get(nomes[1]); 
-            }
-            if (horario == null) {
-                if (inicioOS == null || fimOS == null) return 0;
-                return Duration.between(inicioOS, fimOS).toMinutes();
-            }
-        }
-
-        if (inicioOS == null || fimOS == null) {
-            return 0;
-        }
-
-        long totalMinutosUteis = 0;
-        LocalDateTime cursor = inicioOS; 
-
-        while (cursor.isBefore(fimOS)) {
-            DayOfWeek diaDaSemana = cursor.getDayOfWeek();
-            Turno turnoDoDia = horario.get(diaDaSemana); 
-
-            if (turnoDoDia == null) {
-                cursor = cursor.toLocalDate().plusDays(1).atStartOfDay(); 
-                continue;
-            }
-
-            LocalDateTime inicioTurno = cursor.toLocalDate().atTime(turnoDoDia.inicio);
-            LocalDateTime fimTurno = cursor.toLocalDate().atTime(turnoDoDia.fim);
-            
-            if (turnoDoDia.viraDia()) {
-                if (cursor.toLocalTime().isAfter(turnoDoDia.inicio)) {
-                    fimTurno = cursor.toLocalDate().plusDays(1).atTime(turnoDoDia.fim);
-                } else { 
-                    inicioTurno = cursor.toLocalDate().minusDays(1).atTime(turnoDoDia.inicio);
-                }
-            }
-            
-            LocalDateTime inicioCalculo = cursor.isAfter(inicioTurno) ? cursor : inicioTurno;
-            LocalDateTime fimCalculo = fimOS.isBefore(fimTurno) ? fimOS : fimTurno;
-
-            if (inicioCalculo.isBefore(fimCalculo)) {
-                totalMinutosUteis += Duration.between(inicioCalculo, fimCalculo).toMinutes();
-            }
-
-            if (fimOS.isAfter(fimTurno)) {
-                 cursor = cursor.toLocalDate().plusDays(1).atStartOfDay();
-            } else {
-                cursor = fimOS;
-            }
-        }
-        return totalMinutosUteis;
-    }
+    // ✨ ALTERAÇÃO AQUI: O método 'calcularMinutosUteis' foi removido daqui.
 }
